data/requirements.txt
keras
numpy==1.20
pandas==0.24.2
scikit-learn==0.21.2
tensorflow==2.8.0rc1
protobuf==3.20.0
kafka-python
confluent-kafka